% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{storage_queue}
\alias{storage_queue}
\alias{storage_queue.character}
\alias{storage_queue.queue_endpoint}
\alias{list_storage_queues}
\alias{list_storage_queues.character}
\alias{list_storage_queues.queue_endpoint}
\alias{list_storage_containers.queue_endpoint}
\alias{create_storage_queue}
\alias{create_storage_queue.character}
\alias{create_storage_queue.queue_endpoint}
\alias{create_storage_queue.StorageQueue}
\alias{delete_storage_queue}
\alias{delete_storage_queue.character}
\alias{delete_storage_queue.queue_endpoint}
\alias{delete_storage_queue.StorageQueue}
\title{Message queues}
\usage{
storage_queue(endpoint, ...)

\method{storage_queue}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{storage_queue}{queue_endpoint}(endpoint, name, ...)

list_storage_queues(endpoint, ...)

\method{list_storage_queues}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{list_storage_queues}{queue_endpoint}(endpoint, ...)

\method{list_storage_containers}{queue_endpoint}(endpoint, ...)

create_storage_queue(endpoint, ...)

\method{create_storage_queue}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{create_storage_queue}{queue_endpoint}(endpoint, name, ...)

\method{create_storage_queue}{StorageQueue}(endpoint, ...)

delete_storage_queue(endpoint, ...)

\method{delete_storage_queue}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{delete_storage_queue}{queue_endpoint}(endpoint, name, ...)

\method{delete_storage_queue}{StorageQueue}(endpoint, confirm = TRUE, ...)
}
\arguments{
\item{endpoint}{Either a queue endpoint object as created by \link{storage_endpoint}, or a character string giving the URL of the endpoint.}

\item{...}{Further arguments passed to lower-level functions.}

\item{key, token, sas}{If an endpoint object is not supplied, authentication credentials: either an access key, an Azure Active Directory (AAD) token, or a SAS, in that order of priority.}

\item{api_version}{If an endpoint object is not supplied, the storage API version to use when interacting with the host. Currently defaults to \code{"2019-07-07"}.}

\item{name}{The name of the queue to get, create, or delete.}

\item{confirm}{For deleting a queue, whether to ask for confirmation.}
}
\description{
Get, list, create, or delete queues.
}
