% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{storage_queue}
\alias{storage_queue}
\title{Message queues}
\usage{
storage_queue(endpoint, ...)
}
\arguments{
\item{endpoint}{Either a queue endpoint object as created by \link{storage_endpoint}, or a character string giving the URL of the endpoint.}

\item{...}{Further arguments passed to lower-level functions.}

\item{key, token, sas}{If an endpoint object is not supplied, authentication credentials: either an access key, an Azure Active Directory (AAD) token, or a SAS, in that order of priority.}

\item{api_version}{If an endpoint object is not supplied, the storage API version to use when interacting with the host. Currently defaults to \code{"2019-07-07"}.}

\item{name}{The name of the queue to get, create, or delete.}

\item{confirm}{For deleting a queue, whether to ask for confirmation.}

\item{x}{For the print method, a queue object.}
}
\description{
Get, list, create, or delete queues.
}
