% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{queue}
\alias{queue}
\alias{queue.character}
\alias{queue.queue_endpoint}
\alias{print.queue}
\alias{list_queues}
\alias{list_queues.character}
\alias{list_queues.queue_endpoint}
\alias{list_storage_containers.queue}
\alias{create_queue}
\alias{create_queue.character}
\alias{create_queue.queue}
\alias{create_queue.queue_endpoint}
\alias{delete_queue}
\alias{delete_queue.character}
\alias{delete_queue.queue}
\alias{delete_queue.queue_endpoint}
\title{Operations on a queue endpoint}
\usage{
queue(endpoint, ...)

\method{queue}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{queue}{queue_endpoint}(endpoint, name, ...)

\method{print}{queue}(x, ...)

list_queues(endpoint, ...)

\method{list_queues}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{list_queues}{queue_endpoint}(endpoint, ...)

\method{list_storage_containers}{queue}(endpoint, ...)

create_queue(endpoint, ...)

\method{create_queue}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{create_queue}{queue}(endpoint, ...)

\method{create_queue}{queue_endpoint}(endpoint, name, ...)

delete_queue(endpoint, ...)

\method{delete_queue}{character}(
  endpoint,
  key = NULL,
  token = NULL,
  sas = NULL,
  api_version = getOption("azure_storage_api_version"),
  ...
)

\method{delete_queue}{queue}(endpoint, ...)

\method{delete_queue}{queue_endpoint}(endpoint, name, confirm = TRUE, ...)
}
\arguments{
\item{endpoint}{Either a queue endpoint object as created by \link{storage_endpoint}, or a character string giving the URL of the endpoint.}

\item{...}{Further arguments passed to lower-level functions.}

\item{key, token, sas}{If an endpoint object is not supplied, authentication credentials: either an access key, an Azure Active Directory (AAD) token, or a SAS, in that order of priority.}

\item{api_version}{If an endpoint object is not supplied, the storage API version to use when interacting with the host. Currently defaults to \code{"2019-07-07"}.}

\item{name}{The name of the queue to get, create, or delete.}

\item{x}{For the print method, a queue object.}

\item{confirm}{For deleting a queue, whether to ask for confirmation.}
}
\description{
Get, list, create, or delete queues.
}
