% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{QueueMessage}
\alias{QueueMessage}
\title{R6 class representing a message from an Azure storage queue}
\description{
R6 class representing a message from an Azure storage queue

R6 class representing a message from an Azure storage queue
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{queue}}{The queue this message is from, an object of class \code{\link{StorageQueue}}}

\item{\code{id}}{The message ID.}

\item{\code{insertion_time}}{The message insertion (creation) time.}

\item{\code{expiry_time}}{The message expiration time.}

\item{\code{text}}{The message text.}

\item{\code{receipt}}{A pop receipt. This is populated if the message was retrieved via a \code{get_message} or \code{update} call, and is necessary for deleting or further updating the message.}

\item{\code{next_visible_time}}{The time when this message will be next visible.}

\item{\code{dequeue_count}}{The number of times this message has been read.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QueueMessage$new()}}
\item \href{#method-delete}{\code{QueueMessage$delete()}}
\item \href{#method-update}{\code{QueueMessage$update()}}
\item \href{#method-clone}{\code{QueueMessage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new message object. Rather than calling the \code{initialize} method manually, objects of this class should be created via the methods exposed by the \code{\link{StorageQueue}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueMessage$new(message, queue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{Details about the message.}

\item{\code{queue}}{Object of class `StorageQueue.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Deletes this message from the queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueMessage$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates this message in the queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueMessage$update(visibility_timeout, text = self$text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{visibility_timeout}}{The new visibility timeout (time to when the message will again be visible).}

\item{\code{text}}{Optionally, new message text.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueueMessage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
